+++ src/main/java/org/example/WithdrawalGUI.java
@@ -15,7 +15,7 @@ public class WithdrawalGUI extends AppService implements ActionListener {
     JTextField amountField;
     JLabel messageLabel;
 
-    int amount;
+    double amount;
 
     ResultSet resultSet;
 
@@ -28,7 +28,7 @@ public class WithdrawalGUI extends AppService implements ActionListener {
 
             resultSet = pst.executeQuery();
             if(resultSet.next()){
-                ballance = resultSet.getInt(4);
+                ballance = resultSet.getDouble(4);
                 currency = resultSet.getString(5);
                 withdrawalLimit = resultSet.getInt(6);
             }
@@ -155,14 +155,14 @@ public class WithdrawalGUI extends AppService implements ActionListener {
 
 //This method modify ballance after withdrawal
 
-    public void ballanceUpdate(String uname, int ballance){
+    public void ballanceUpdate(String uname, double ballance){
         String SQL = "UPDATE atmusers "
                 + "SET ballance = ? "
                 + "WHERE username = ?";
 
         try (PreparedStatement pstmt = con.prepareStatement(SQL))  {
 
-            pstmt.setInt(1,ballance);
+            pstmt.setDouble(1,ballance);
             pstmt.setString(2,uname);
 
             pstmt.executeUpdate();
@@ -182,7 +182,7 @@ public class WithdrawalGUI extends AppService implements ActionListener {
         }
         else if (e.getSource() == withdrawalButton){
             try {
-                amount = Integer.parseInt(amountField.getText());
+                amount = Double.parseDouble(amountField.getText());
                 if(ballance-amount < 0){
                     messageLabel.setText("Insufficient funds, check your ballance!");
                     messageLabel.setVisible(true);
